+++ package/spec/styles.spec.tsx
@@ -6,6 +6,14 @@ describe("styles", () => {
         b.styleDef([userSelectNone, { fill: "red" }]);
     });
 
+    it("can declare keyframes", () => {
+        let red2green = b.keyframesDef({ from: { color: "red" }, to: { color: "green" } });
+        b.init(() => <div style={{ animation: red2green("2s") }}>Hello</div>);
+        b.syncUpdate();
+        expect(document.head.innerHTML).toContain("@keyframes ");
+        expect(document.head.innerHTML).toContain("from {color:red}");
+    });
+
     it("can apply variable style", () => {
         let a = 1;
         b.style(<div />, [a > 1 || { display: "flex" }, [{ alignSelf: "baseline" }]]);
@@ -13,14 +21,14 @@ describe("styles", () => {
 
     it("!!!temporary allowed!!! 0 in styles", () => {
         let content: b.IBobrilChildren = 0;
-        b.init(() => <div style={ content && { color:"red" } }>I must not be red</div>);
+        b.init(() => <div style={content && { color: "red" }}>I must not be red</div>);
         b.syncUpdate();
         expect(document.body.innerHTML).not.toContain("color: red");
     });
 
     it("correct check for content existence in styling", () => {
         let content: b.IBobrilChildren = 42;
-        b.init(() => <div style={ content !== undefined && { color:"red" } }>I must be red {content}</div>);
+        b.init(() => <div style={content !== undefined && { color: "red" }}>I must be red {content}</div>);
         b.syncUpdate();
         expect(document.body.innerHTML).toContain("color: red");
         expect(document.body.innerHTML).toContain("I must be red 42");